requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
groq==0.4.1
sqlalchemy==2.0.25
python-dotenv==1.0.0
redis==5.0.1
pdfplumber==0.10.3
librosa==0.10.0
soundfile==0.12.1
numpy==1.24.0
scipy==1.12.0
webrtcvad==2.0.10
pyttsx3==2.90
gTTS==2.3.2
pymongo==4.3.3
